vbus-bridge: add support for legacy interrupt generation
Some platforms (e.g. Windows) have trouble supporting MSI based interrupts,
so lets add support for the legacy INTx pin-based interrupts as well.

We implement this change purely in userspace by terminating an eventfd
in a qemu based thread, and passing the event on as a standard
qemu_set_irq operation.  This way the kernel still sees the
injection path as an eventfd, just like it does when we use msi/irqfd.

This was tested with both a MSI and INTx based linux guest, and passes
traffic properly.

Known limitations: resetting a guest is not fully supported, but it was
already broken in vbus already. We will address reset as a whole in a
future update.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>