reset the vbus bridge on NEGOTIATE
We want to support the notion of reloading the bridge driver in the
guest without rebooting the VM.  However, we presently have no way
to reset the bridge device once it is configured.  This patch
adds an implicit reset capability tied to the NEGOTIATE feature.
Since NEGOTIATE is only invoked as a driver initializes, this gives
us a good indication that it is safe to "start over" and reset
the world.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>